Add the new arrival and departure fields. HUGE CHANGE
The new arrival and departure fields are of a new type with a date and
a time picker. They will also be stored in a new field. THIS IS NOT
BACKWARD COMPATIBLE AND A STRAIGHT UPGRADE WILL MEAN NO MORE DEPARTURE
OR ARRIVALS WILL DISPLAY IN A THEME. A convertor will be added to the
upgrade script that will pull existing departure and arrival values
out and put them into the new format.